# Optimize yolo model postprocessing for python detection demo