Optimize yolo postprocessing for detection demo, by using flatten operation and apply sigmoid function to all boxes at the same time.
**Speed up yoloV4**:
1. For YoloV4: `2.4 -> 3.7 fps` It's x1.55.
2. For YoloV4-tiny: `25 -> 50+ fps` It's x2.

The calculations were carried out on the CPU `Intel® Core™ i7-8700 CPU` and on video with resolution `1280 × 720`. `nstreams = 4, nrequests = 5, nthreads = auto` 

**Speed up yolo family**:
1. For `yolo-v3-tf`: `7.5 -> 8.4 fps` , **x1.12**
2. For `yolo-v3-tiny-tf`: `32 -> 62.8 fps`, **x1.96**
3. For `yolo-v2-tf`: `9 -> 9.1 fps`, x1
4. For `yolo-v2-tiny-tf`: `60 -> 70 fps`, **x1.16**
5. For `yolo-v2-tiny-ava-0001`: `51.6 -> 56.5 fps`, x1.1
6. For `yolo-v2-tiny-ava-sparse-30-0001`: `51.3 -> 57 fps`, x1.11
7. For `yolo-v2-tiny-ava-sparse-60-0001`: `52.5 -> 57 fps`, x1.085
8. For `yolo-v2-ava-0001`: `17.5 -> 17.8 fps`, x1
9. For `yolo-v2-ava-sparse-35-0001`: `17.5 -> 17.8 fps`, x1
10. For `yolo-v2-ava-sparse-70-0001`: `17.5 -> 17.8 fps`, x1
11. For `yolo-v2-tiny-vehicle-detection-0001`: `64.5 -> 68.2 fps`, x1.05
12. For `yolo-v1-tiny-tf`: `54.8 -> 57.5 fps`, x1.05
13. For `person-vehicle-bike-detection-crossroad-yolov3-1020`: `7.5 -> 8.4 fps` , **x1.12**
14. For `mobilefacedet-v1-mxnet`: `47.5 -> 99.5 fps`, **x2.1**

The calculations were carried out on the CPU `Intel® Core™ i7-8700 CPU` and on images from coco dataset. `nstreams = 4, nrequests = 5, nthreads = auto` 

Now Yolo model support only **Multiple_labels mode**, as at c++ demo