@vladimir-dudnik 2021-09-07
@akorobeinikov would it make sense to do the same on AC side, so we can estimate if this optimization affects accuracy?