@akorobeinikov 2021-09-07
> would it make sense to do the same on AC side, so we can estimate if this optimization affects accuracy?

In my opinion, it is don't make sense, because this optimizations haven't any new operations, it only reorder steps of postprocessing for speed up, but new version is equivalent to old version and optimization can affect accuracy only if there are  bugs in code. Also I checked accuracy of boxes when tested performance on single images and accuracy is the same (this is not indicative, but at least something). We can do the same on AC side for optimization, but I think it is don't make sense to change AC yolo structure for little optimization, because AC don't oriented on speed.