@vladimir-dudnik 2021-09-07
> > would it make sense to do the same on AC side, so we can estimate if this optimization affects accuracy?
> 
> In my opinion, it is don't make sense, because this optimizations haven't any new operations,
That is exactly what might be tested with AC. And additionally, improving speed for AC validation is important too.